% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_indicator.R
\name{generate_map_legend}
\alias{generate_map_legend}
\title{Generate legend for a BCU indicator map}
\usage{
generate_map_legend(indicator)
}
\arguments{
\item{indicator}{Indicator being mapped}
}
\value{
A ggplot2 object
}
\description{
Generate a ggplot2 legend for a BCU indicator map. The ggplot2 legend is more flexible than the tmap legend, so a ggplot2 legend is produced for every map from \code{map_indicator}.
}
\examples{
generate_map_legend("grav_NC")
}
