% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report_card.R
\name{generate_report_cards}
\alias{generate_report_cards}
\title{Generate Reef Pressures report cards}
\usage{
generate_report_cards(bcu = "all", open = TRUE, quiet = FALSE)
}
\arguments{
\item{bcu}{BCU(s) to generate report cards for. Defaults to "all", to generate for all BCUs.}

\item{open}{Whether to open the PDF once the report card is generated. Defaults to TRUE.}

\item{quiet}{Whether to suppress messages about which report card is being produced and where it is saved. Defaults to FALSE.}
}
\value{
A set of pdf report cards saved in \code{report-cards/}
}
\description{
Generate Reef Pressures report cards
}
