% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report_card.R
\name{generate_single_report_card}
\alias{generate_single_report_card}
\title{Generate a single reef pressures report card}
\usage{
generate_single_report_card(bcu, open = TRUE, quiet = FALSE, standalone = TRUE)
}
\arguments{
\item{bcu}{A single BCU to generate report cards for}

\item{open}{Whether to open the PDF once the report card is generated. Defaults to TRUE}

\item{quiet}{Whether to suppress messages about which report card is being produced and where it is saved. Defaults to FALSE}

\item{standalone}{Whether the report card is being produced standalone (i.e., only one report card is being produced) - defaults to TRUE. This controls whether the libraries and data are loaded when the function is called.}
}
\value{
A pdf report card saved in \code{analysis/report-cards/}
}
\description{
Generate a single reef pressures report card
}
\examples{
generate_single_report_card("Eel - NGBR")
}
