% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_left_right_tiles.R
\name{get_left_right_tiles}
\alias{get_left_right_tiles}
\title{Get left and right tiles}
\usage{
get_left_right_tiles(bcus, bcu_name, bcu_ext, tile_type = "bing", zoom = NULL)
}
\arguments{
\item{bcus}{Indicator data for BCUs, from \code{data/bcus_list.rds}}

\item{bcu_name}{Name of BCU that tiles are being pulled for}

\item{bcu_ext}{Extension factor of the BCU's bounding box, from \code{data/bcus_ext.rds}}

\item{tile_type}{Type of OSM tile to get. Defaults to "bing".}

\item{zoom}{Tile zoom. Defaults to NULL, in which case it is determined automatically}
}
\value{
A named list ("left" and "right") of OSM tiles.
}
\description{
Get left and right OSM tiles for a given BCU - necessary for BCUs that cross the 180th meridian line (i.e., Fiji BCUs), as \code{tmaptools::read_osm} uses a fixed projection that cannot be changed and so cuts off.
}
\examples{
bcu_name <- "Vatu-i-Ra"
bcus <- readRDS(here::here("data", "bcus_list.rds"))
bcu_ext <- readRDS(here::here("data", "bcus_ext.rds"))[[bcu_name]]

get_left_right_tiles(bcus, bcu_name, bcu_ext)
}
