% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_indicator.R
\name{map_indicator}
\alias{map_indicator}
\title{Map BCU Indicator}
\usage{
map_indicator(bcu, tile, indicator = NA, indicator_title = indicator, bcu_ext)
}
\arguments{
\item{bcu}{Indicator data for a single BCU, from \code{data/bcus_list.rds}}

\item{tile}{Bing tile for a single BCU, from \code{data/bcus_tiles_indicator.rds}}

\item{indicator}{Indicator being mapped}

\item{indicator_title}{Longer-form title for indicator being mapped, used in creating the map title. Defaults to \code{indicator}}

\item{bcu_ext}{Extension factor of the BCU's bounding box, from \code{data/bcus_ext.rds}}
}
\value{
A \code{tmap} object
}
\description{
Produce a satellite map for a BCU, along with the values for a supplied indicator.
}
\examples{
bcu_name <- "Aceh"
bcu <- readRDS(here::here("data", "bcus_list.rds"))[[bcu_name]]
tile <- readRDS(here::here("data", "bcus_tiles_indicator.rds"))[[bcu_name]]
bcu_ext <- readRDS(here::here("data", "bcus_ext.rds"))[[bcu_name]]

map_indicator(bcu, tile,
  indicator = "grav_NC",
  indicator_title = "Fishing: Market Pressure", bcu_ext
)
}
