% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indicator.R
\name{plot_indicator_dumbbell}
\alias{plot_indicator_dumbbell}
\title{Plot indicator values}
\usage{
plot_indicator_dumbbell(bcus_top_pressures, bcus_regions, bcu_name, indicator)
}
\arguments{
\item{bcus_top_pressures}{Top pressures for all BCUs, from \code{data/bcus_top_pressures.rds}}

\item{bcus_regions}{Data containing which regions BCUs belong to, from \code{data/bcus_regions.rds}}

\item{bcu_name}{Name of the BCU being plotted}

\item{indicator}{Indicator being plotted}
}
\value{

}
\description{
Produce a ggplot2 plot of a BCU's indicator values, along with the global median for that indicator and the medians of other BCUs in that subregion.
}
\examples{
bcus_top_pressures <- readRDS(here::here("data", "bcus_top_pressures.rds"))
bcus_regions <- readRDS(here::here("data", "bcus_regions.rds"))

plot_indicator_dumbbell(bcus_top_pressures, bcus_regions, "Aceh", "grav_NC")
}
