% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pressure_ranking_percentiles.R
\name{plot_pressure_ranking_percentiles}
\alias{plot_pressure_ranking_percentiles}
\title{Plot top pressures percentiles}
\usage{
plot_pressure_ranking_percentiles(bcu_percentiles, bcu_top_pressures)
}
\arguments{
\item{bcu_percentiles}{Percentiles for a single BCU, from \code{data/bcus_percentiles.rds}}

\item{bcu_top_pressures}{Top pressures for a single BCU, from \code{data/bcus_top_pressures.rds}}
}
\value{
A ggplot2 object
}
\description{
Produce a ggplot2 plot showing the distribution of a BCU's pixels percentiles, relative to all reef pixels, and the percentile of the BCU's average, for all indicators (except number of ports, and excluding climate scores). The indicators are ordered in descending order by the value of the indicator average's percentile.
}
\examples{
bcu_name <- "Aceh"
bcu_percentiles <- readRDS(here::here("data", "bcus_percentiles.rds"))[[bcu_name]]
bcu_top_pressures <- readRDS(here::here("data", "bcus_top_pressures.rds"))[[bcu_name]]

plot_pressure_ranking_percentiles(bcu_percentiles, bcu_top_pressures)
}
