% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_bcus.R
\name{recode_bcus}
\alias{recode_bcus}
\title{Recode BCUs}
\usage{
recode_bcus(bcu)
}
\arguments{
\item{bcu}{A vector of BCU names}
}
\value{
A vector of BCU names - cleaned if needed, otherwise just returns the original BCU name.
}
\description{
Clean BCU names for consistency.
}
\examples{
recode_bcus("Eel - NGBR")
# [1] "Eel - Northern GBR"
recode_bcus("North Red Sea-Egypt II")
# [1] "North Red Sea - Egypt II"
recode_bcus("Bahamas")
# [1] Bahamas
}
