% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_bbox.R
\name{square_bbox}
\alias{square_bbox}
\title{Generate a square bounding box}
\usage{
square_bbox(x, ext = 1.1, output = "bbox")
}
\arguments{
\item{x}{Data to create bounding box for}

\item{ext}{Extension factor of the bounding box (relative to the data in \code{x}). A value greater than 1 means the bounding box will be larger than the bounds of the data, and a value less than 1 means the bounding box will be smaller.}

\item{output}{Type of output. Defaults to "bbox"}
}
\value{
A bounding box
}
\description{
Generate a square bounding box
}
\examples{
bcu <- readRDS(here::here("data", "bcus_list.rds"))[["Singapore"]]
square_bbox(bcu)
}
